requirements.txt
dj-database-url==0.5.0
Django==2.0.7
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
psycopg2==2.7.5
pytz==2018.5
